+++ include/history.h
@@ -8,7 +8,7 @@ namespace gbolt {
 
 class History {
  public:
-  History(size_t max_edges, size_t max_vertice) : edge_size_(0) {
+  History(int max_edges, int max_vertice) : edge_size_(0) {
     edges_ = new ConstEdgePointer[max_edges + 1];
     has_edges_ = new bool[max_edges + 1]();
     has_vertice_ = new bool[max_vertice + 1]();
@@ -20,15 +20,15 @@ class History {
 
   void build_vertice(const struct prev_dfs_t &start, const Graph &graph);
 
-  bool has_edges(size_t index) const {
+  bool has_edges(int index) const {
     return has_edges_[index];
   }
 
-  bool has_vertice(size_t index) const {
+  bool has_vertice(int index) const {
     return has_vertice_[index];
   }
 
-  const struct edge_t *get_p_edge(size_t index) const {
+  const struct edge_t *get_p_edge(int index) const {
     return edges_[edge_size_ - index - 1];
   }
 
@@ -43,7 +43,7 @@ class History {
   ConstEdgePointer *edges_;
   bool *has_edges_;
   bool *has_vertice_;
-  size_t edge_size_;
+  int edge_size_;
 };
 
 }  // namespace gbolt
